chore: convert agent spec to vitest
